refactor(git): split branch commands and refresh baseline
