+++ docs/architecture/large-file-baseline.md
@@ -1,17 +1,11 @@
 # Large File Hard-Debt Baseline
 
-- Generated at: 2026-04-22T17:01:09.309Z
+- Generated at: 2026-04-22T17:37:49.723Z
 - Scope: fail
 - Policy version: 2026-04-22.policy-v2
-- Count: 7
+- Count: 1
 
 | File | Lines | Type | Policy | Priority | Warn | Fail | Severity | Status | Baseline | Delta |
 |---|---:|---|---|---|---:|---:|---|---|---:|---:|
-| `src/styles/settings.part1.css` | 2993 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
 | `src-tauri/src/engine/commands.rs` | 2976 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
-| `src/styles/composer.part1.css` | 2941 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
-| `src/styles/settings.part2.css` | 2860 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
-| `src/styles/git-history.part1.css` | 2818 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
-| `src-tauri/src/runtime/mod.rs` | 2655 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
-| `src-tauri/src/git/commands.rs` | 2619 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
