+++ docs/architecture/large-file-near-threshold-watchlist.md
@@ -1,38 +1,34 @@
 # Large File Near-Threshold Watchlist
 
-- Generated at: 2026-04-22T17:01:09.717Z
+- Generated at: 2026-04-22T17:37:50.166Z
 - Scope: warn
 - Policy version: 2026-04-22.policy-v2
-- Count: 31
+- Count: 27
 
 | File | Lines | Type | Policy | Priority | Warn | Fail | Severity | Status | Baseline | Delta |
 |---|---:|---|---|---|---:|---:|---|---|---:|---:|
 | `src/features/threads/hooks/useThreadActions.test.tsx` | 2995 | ts/tsx | test-files | P2 | 2600 | 3000 | warn | watch |  |  |
-| `src/styles/settings.part1.css` | 2993 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
 | `src-tauri/src/engine/commands.rs` | 2976 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
 | `src/features/app/hooks/useAppServerEvents.test.tsx` | 2966 | ts/tsx | test-files | P2 | 2600 | 3000 | warn | watch |  |  |
 | `src/features/threads/hooks/useThreadsReducer.test.ts` | 2961 | ts/tsx | test-files | P2 | 2600 | 3000 | warn | watch |  |  |
-| `src/styles/composer.part1.css` | 2941 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
 | `src-tauri/src/local_usage.rs` | 2929 | rust | default-source | P1 | 2600 | 3000 | warn | watch |  |  |
 | `src/features/messages/components/Messages.test.tsx` | 2926 | ts/tsx | test-files | P2 | 2600 | 3000 | warn | watch |  |  |
 | `src-tauri/src/bin/cc_gui_daemon.rs` | 2882 | rust | default-source | P1 | 2600 | 3000 | warn | watch |  |  |
-| `src/styles/settings.part2.css` | 2860 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
 | `src/i18n/locales/en.part1.ts` | 2835 | i18n | i18n | P2 | 2600 | 3000 | warn | watch |  |  |
-| `src/styles/git-history.part1.css` | 2818 | css | styles | P1 | 2200 | 2800 | fail | captured |  |  |
 | `src/features/git/components/GitDiffPanel.tsx` | 2807 | ts/tsx | default-source | P1 | 2600 | 3000 | warn | watch |  |  |
 | `src/styles/git-history.part2.css` | 2776 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
 | `src/features/git-history/components/git-history-panel/components/GitHistoryPanelImpl.tsx` | 2773 | ts/tsx | feature-hotpath | P1 | 2400 | 2800 | warn | watch |  |  |
 | `src/i18n/locales/zh.part1.ts` | 2743 | i18n | i18n | P2 | 2600 | 3000 | warn | watch |  |  |
 | `src/features/threads/hooks/useThreadsReducer.ts` | 2723 | ts/tsx | feature-hotpath | P1 | 2400 | 2800 | warn | watch |  |  |
 | `src-tauri/src/bin/cc_gui_daemon/daemon_state.rs` | 2713 | rust | default-source | P1 | 2600 | 3000 | warn | watch |  |  |
-| `src-tauri/src/runtime/mod.rs` | 2655 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
 | `src/styles/sidebar.css` | 2640 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
-| `src-tauri/src/git/commands.rs` | 2619 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | fail | captured |  |  |
 | `src/styles/spec-hub.css` | 2609 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
 | `src/features/settings/components/SettingsView.tsx` | 2607 | ts/tsx | feature-hotpath | P1 | 2400 | 2800 | warn | watch |  |  |
+| `src/styles/git-history.part1.css` | 2592 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
 | `src-tauri/src/codex/mod.rs` | 2574 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
 | `src/services/tauri.ts` | 2567 | ts/tsx | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
 | `src/app-shell.tsx` | 2509 | ts/tsx | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
+| `src-tauri/src/runtime/mod.rs` | 2372 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
 | `src/styles/tool-blocks.css` | 2265 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
 | `src/styles/file-view-panel.css` | 2237 | css | styles | P1 | 2200 | 2800 | warn | watch |  |  |
 | `src-tauri/src/backend/app_server.rs` | 2235 | rust | bridge-runtime-critical | P0 | 2200 | 2600 | warn | watch |  |  |
